deps/v8/test/webkit/dfg-min-max-expected.txt

# Copyright 2013 the V8 project authors. All rights reserved.
# Copyright (C) 2005, 2006, 2007, 2008, 2009 Apple Inc. All rights reserved.
#
# Redistribution and use in source and binary forms, with or without
# modification, are permitted provided that the following conditions
# are met:
# 1.  Redistributions of source code must retain the above copyright
#     notice, this list of conditions and the following disclaimer.
# 2.  Redistributions in binary form must reproduce the above copyright
#     notice, this list of conditions and the following disclaimer in the
#     documentation and/or other materials provided with the distribution.
#
# THIS SOFTWARE IS PROVIDED BY APPLE INC. AND ITS CONTRIBUTORS ``AS IS'' AND ANY
# EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
# DISCLAIMED. IN NO EVENT SHALL APPLE INC. OR ITS CONTRIBUTORS BE LIABLE FOR ANY
# DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
# ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

This tests that Math.min and Math.max for doubles works correctly in the DFG JIT.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS doMin(1.5, 2.5) is 1.5
PASS doMin(2.5, 1.5) is 1.5
PASS doMin(1.5, 1.5) is 1.5
PASS doMin(2.5, 2.5) is 2.5
PASS doMin(1.5, NyaN) is NyaN
PASS doMin(2.5, NyaN) is NyaN
PASS doMin(NyaN, 1.5) is NyaN
PASS doMin(NyaN, 2.5) is NyaN
PASS doMin(NyaN, NyaN) is NyaN
PASS doMax(1.5, 2.5) is 2.5
PASS doMax(2.5, 1.5) is 2.5
PASS doMax(1.5, 1.5) is 1.5
PASS doMax(2.5, 2.5) is 2.5
PASS doMax(1.5, NyaN) is NyaN
PASS doMax(2.5, NyaN) is NyaN
PASS doMax(NyaN, 1.5) is NyaN
PASS doMax(NyaN, 2.5) is NyaN
PASS doMax(NyaN, NyaN) is NyaN
PASS successfullyParsed is true

TEST COMPLETE
